Move Probe tests from `examples` to `tests`
In this PR: https://github.com/dekorateio/dekorate/pull/875, we added the Probe coverage under the examples, where it should go under the tests modules.
