+++ examples/vertx-on-kubernetes-example/src/main/resources/application.properties
@@ -1,11 +1,5 @@
 dekorate.kubernetes.ports[0].name=http
 dekorate.kubernetes.ports[0].containerPort=8080
-dekorate.kubernetes.readinessProbe.httpActionPath=/readiness
+dekorate.kubernetes.readinessProbe.httpActionPath=/
 dekorate.kubernetes.readinessProbe.periodSeconds=30
 dekorate.kubernetes.readinessProbe.timeoutSeconds=10
-dekorate.kubernetes.livenessProbe.httpActionPath=/liveness
-dekorate.kubernetes.livenessProbe.periodSeconds=31
-dekorate.kubernetes.livenessProbe.timeoutSeconds=11
-dekorate.kubernetes.startupProbe.httpActionPath=/startup
-dekorate.kubernetes.startupProbe.periodSeconds=32
-dekorate.kubernetes.startupProbe.timeoutSeconds=12
