+++ examples/vertx-on-openshift-example/src/main/resources/application.properties
@@ -1,11 +1,5 @@
 dekorate.openshift.ports[0].name=http
 dekorate.openshift.ports[0].containerPort=8080
-dekorate.openshift.readinessProbe.httpActionPath=/readiness
+dekorate.openshift.readinessProbe.httpActionPath=/
 dekorate.openshift.readinessProbe.periodSeconds=30
 dekorate.openshift.readinessProbe.timeoutSeconds=10
-dekorate.openshift.livenessProbe.httpActionPath=/liveness
-dekorate.openshift.livenessProbe.periodSeconds=31
-dekorate.openshift.livenessProbe.timeoutSeconds=11
-dekorate.openshift.startupProbe.httpActionPath=/startup
-dekorate.openshift.startupProbe.periodSeconds=32
-dekorate.openshift.startupProbe.timeoutSeconds=12
